Former-commit-id: 703a98357675d2e0c14a44392ff7c23af8d49bfc

+++ src/runtime/HalideRuntimeMetal.h
@@ -41,17 +41,16 @@ extern int halide_metal_run(void *user_context,
  * routine is called. This call can fail due to running out of memory
  * or being passed an invalid buffer. The device and host dirty bits
  * are left unmodified. */
-extern int halide_metal_wrap_buffer(void *user_context, struct halide_buffer_t *buf, uintptr_t buffer);
+extern int halide_metal_wrap_buffer(void *user_context, struct halide_buffer_t *buf, uint64_t buffer);
 
-/** Disconnect a halide_buffer_t from the memory it was previously wrapped
- * around. Should only be called for a halide_buffer_t that
+/** Disconnect a halide_buffer_t from the memory it was previously
+ * wrapped around. Should only be called for a halide_buffer_t that
  * halide_metal_wrap_buffer was previously called on. Frees any
- * storage associated with the binding of the halide_buffer_t and the buffer,
- * but does not free the MTLBuffer. The previously wrapped
- * buffer is returned. The dev field of the halide_buffer_t will be NULL on
- * return.
+ * storage associated with the binding of the halide_buffer_t and the
+ * buffer, but does not free the MTLBuffer. The dev field of the
+ * halide_buffer_t will be NULL on return.
  */
-extern uintptr_t halide_metal_detach_buffer(void *user_context, struct halide_buffer_t *buf);
+extern int halide_metal_detach_buffer(void *user_context, struct halide_buffer_t *buf);
 
 /** Return the underlying MTLBuffer for a halide_buffer_t. This buffer must be
  * valid on an Metal device, or not have any associated device
